Refactor service folder
