+++ internal/core/service/xkcd_test.go
